+++ application/messages/views.py
@@ -83,6 +83,10 @@ def messages_edit(message_id):
         return render_template("messages/edit.html", form = form, message_id = m.id)
     
     edit = Message.query.filter_by(id=message_id).first()
+
+    if not edit.account_id == current_user.id:
+        return redirect(url_for("messages_index"))
+
     if edit.messagetext == request.form['messagetext'] and edit.category_id == request.form['category_id']:
         return redirect(url_for("messages_index"))
     else:
@@ -95,14 +99,17 @@ def messages_edit(message_id):
 @app.route("/messages/delete/<message_id>/", methods=["POST"])
 @login_required
 def messages_delete(message_id):
+    message = Message.query.filter_by(id=message_id).first()
+    if not message.account_id == current_user.id and not current_user.id == 1:
+        return redirect(url_for("messages_index"))
+
     answersToDelete = Answer.query.filter_by(message_id=message_id)
     answersToDelete.delete()
 
     likesToDelete = Like.query.filter_by(message_id=message_id)
     likesToDelete.delete()
     
-    db.session.delete(Message.query.filter_by(id=message_id).first())
-    
+    db.session.delete(message)
     db.session.commit()
 
     return redirect(url_for("messages_index"))
